Fixed a bug creating the first deck of cards.
Since the first deck of cards creats the entire deck object, it needed to be set using setItem rather than mergeItem. Handling for each of these methods and its returns was added in this update. In order to avoid duplicate code, the initial AsyncStorage handling was moved to an external method, and createDeck was adjusted to use it (and work from its returned promise).